[#222] feat: 멤버 목록 cell UI 다듬기
