+++ lib/presentation/group_post/view/group_post_view_widget.dart
@@ -904,66 +904,3 @@ class ConfirmPostContentWidget extends StatelessWidget {
     }
   }
 }
-
-class MemberListCellWidget extends StatelessWidget {
-  const MemberListCellWidget({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Container(
-          decoration: const BoxDecoration(
-            shape: BoxShape.circle,
-            color: CustomColors.whGrey,
-          ),
-          width: 60,
-          height: 60,
-          child: Image.network(
-            fit: BoxFit.cover,
-            'https://health.chosun.com/site/data/img_dir/2023/07/17/2023071701753_0.jpg',
-          ),
-          clipBehavior: Clip.hardEdge,
-        ),
-        SizedBox(
-          width: 20.0,
-        ),
-        Expanded(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                '박성민',
-                style: TextStyle(
-                  color: CustomColors.whWhite,
-                  fontSize: 18.0,
-                  fontWeight: FontWeight.w600,
-                  height: 1.0,
-                ),
-              ),
-              Text(
-                '6개의 목표 공유중',
-                style: TextStyle(
-                  color: CustomColors.whWhite,
-                  fontSize: 14.0,
-                  fontWeight: FontWeight.w300,
-                ),
-              )
-            ],
-          ),
-        ),
-        Text(
-          '67%',
-          style: TextStyle(
-            color: CustomColors.whWhite,
-            fontSize: 20.0,
-            fontWeight: FontWeight.w600,
-          ),
-        ),
-      ],
-    );
-  }
-}
